share MockKernel and fixtures across tests
